requirements.txt
allure-pytest==2.5.3
Appium-Python-Client==0.28
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bidict==0.17.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
configparser==3.5.0
crypto==1.4.1
cryptography==2.3.1
cssselect==1.0.3
ddt==1.2.0
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.0.2
get==1.0.3
gevent==1.3.7
greenlet==0.4.15
idna==2.7
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
locustio==0.9.0
lxml==4.2.5
MarkupSafe==1.0
more-itertools==4.3.0
msgpack==0.5.6
Naked==0.1.31
namedlist==1.7
numpy==1.15.4
opencv-python==3.4.4.19
openpyxl==2.5.9
Pillow==5.3.0
pluggy==0.8.0
post==1.0.2
public==2.0.1
py==1.7.0
pycparser==2.19
PyMySQL==0.9.2
pyquery==1.4.0
pytesseract==0.2.5
pytest==3.9.2
pytest-allure-adaptor==1.7.10
PyYAML==3.13
pyzmq==17.1.2
query-string==1.0.2
request==1.0.2
requests==2.20.0
selenium==3.14.1
shellescape==3.4.1
six==1.11.0
termcolor==1.1.0
uiautomator==0.3.6
urllib3==1.24
websocket-client==0.53.0
Werkzeug==0.14.1
xlrd==1.1.0
